backend/requirements.txt
python-dotenv==1.0.1
openai==1.52.0
uagents==0.16.2